+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Luau analyze now exits with code 0 if there are no reported errors (all errors are ignored)
+
 ## [1.14.3] - 2022-12-10
 
 ### Changed